+++ rmp-serde/CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - More `decode::Error` conversions.
 - Support for serializing and deserializing 128-bit values in serde.
 - Support for serializing sequences and maps with unknown length, that enables the use of `#[serde(flatten)]` attribute (#196).
+- Depth limit is now enforced for `Deserializer`.
 
 ### Changed:
 - (Breaking) Serialize newtype structs by serializing its inner type without wrapping into a tuple.